// src/typings/api.ts
export interface ApiAddOnProductLink {
    productId: number
    productName: string
}
// ...
export interface ApiSubscription {
    simCards: {
        main: ApiSimCard,
        twin: ApiSimCard[],
        data: ApiSimCard[]
    }
    addOnProducts: ApiSubscriptionAddOnProduct[]
}

export interface MobileDeviceInformation {
    vendor: string
    model: string
    imei: string
}

export interface SimDetails {
    iccId: string
    imsi: string
    network: '3G' | '4G',
    esimDetails?: {
        status?: string
        isEsim: boolean
    }
}

export interface BarringStatus {
    barred: boolean
    status: 'OPEN' | 'ADDING' | 'BARRED' | 'REMOVING'
}
export interface ApiSimCard {
    inventoryProductId: string
    msisdn: string
    mobileDeviceInformation?: MobileDeviceInformation
    simDetails: SimDetails
    barringStatus: BarringStatus
}

export interface ApiSubscriptionAddOnProduct extends ApiAddOnProductLink {
    orders: ApiSubscriptionOrder[]
}

export interface ApiSubscriptionOrder {
    inventoryProductId: string
    parentInventoryProductId?: string
    status: 'NO_CHANGE' | 'NEW_OPEN' | 'CHANGE_ALTER' | 'REMOVE_CANCEL'
    originalFromDate: string
    validFrom: string
    validTo?: string
    parameters?: ApiOrderParameter[]
}

export interface ApiOrderParameter {
    parameterId: number
    parameterValue: string
}